requirements.txt
appdirs==1.4.3
dnspython==1.15.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-etcd==0.4.5
six==1.10.0
urllib3==1.20
